requirements.txt
aioredis==1.3.1
amqp==5.0.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==0.8.0
backcall==0.2.0
billiard==3.6.3.0
blinker==1.4
celery==5.0.2
certifi==2020.6.20
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
Click==7.0
click-didyoumean==0.0.3
click-repl==0.1.6
cloud-init==20.3
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cvxopt==1.2.5
cycler==0.10.0
daphne==2.5.0
dbus-python==1.2.16
decorator==4.4.2
deprecation==2.1.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.2
django-celery-beat==2.1.0
django-timezone-field==4.0
entrypoints==0.3
graphviz==0.14.2
hiredis==1.1.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
intervaltree==3.1.0
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.10.1
joblib==0.17.0
jsonpatch==1.22
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.3.0
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.6.1
MarkupSafe==1.1.0
matplotlib==3.3.2
more-itertools==4.2.0
mpmath==1.1.0
msgpack==1.0.0
netifaces==0.10.4
networkx==2.5
numpy==1.19.3
oauthlib==3.1.0
packaging==20.4
pandas==1.1.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pm4py==2.0.0
pm4pycvxopt==0.0.10
prompt-toolkit==3.0.8
ptyprocess==0.6.0
PuLP==2.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydotplus==2.0.2
Pygments==2.7.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
pytz==2020.1
pyvis==0.1.8.2
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
scikit-learn==0.23.2
scipy==1.5.3
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sortedcontainers==2.2.2
sqlparse==0.3.1
ssh-import-id==5.10
StringDist==1.0.9
sympy==1.6.2
systemd-python==234
threadpoolctl==2.1.0
tqdm==4.51.0
traitlets==5.0.5
Twisted==18.9.0
txaio==20.4.1
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.5
zipp==1.0.0
zope.interface==4.7.1
